requirements.txt
Pillow==10.0.0
scikit-learn==1.2.0
numpy==1.24.0